Eric Sunshine eliminated numerous compilation warnings throughout the project.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@13799 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/loader/dds/dds.cpp
@@ -212,10 +212,11 @@ bool Loader::ReadHeader ()
     else
       add = MAX(add/4, 16);
   }
-  if ((positions[GetMipmapCount()] - positions[0]) + add > sourcelen)
+  if ((size_t)(positions[GetMipmapCount()] - positions[0] + add) > sourcelen)
   {
-    printf ("DDS Image too small Needs:%u Has: %d.\n",
-	(positions[GetMipmapCount()] - positions[0])+add, sourcelen);
+    printf ("DDS Image too small Needs:%u Has: %lu.\n",
+	(unsigned int)(positions[GetMipmapCount()] - positions[0] + add),
+	(unsigned long)sourcelen);
     return false;
   }
